+++ src/style.rs
@@ -528,3 +528,82 @@ impl<'s> From<StyledString> for StyledCow<'s> {
         StyledCow::new(s.s, s.style)
     }
 }
+
+/// A style for a line, used in styling borders and shapes.
+///
+/// The style consists of:
+/// - the line thickness in millimeters (defaults to 0.1)
+/// - the color of the line, see [`Color`][] (defaults to black)
+///
+/// Note that a line thickness of 0.0 does not make the line disappear, but rather makes it appear
+/// 1px wide across all devices and resolutions.
+///
+/// [`Color`]: enum.Color.html
+#[derive(Clone, Copy, Debug, PartialEq)]
+pub struct LineStyle {
+    thickness: Mm,
+    color: Color,
+}
+
+impl Default for LineStyle {
+    fn default() -> LineStyle {
+        LineStyle {
+            thickness: Mm::from(0.1),
+            color: Color::Rgb(0, 0, 0),
+        }
+    }
+}
+
+impl From<Color> for LineStyle {
+    fn from(color: Color) -> LineStyle {
+        LineStyle {
+            color,
+            ..LineStyle::default()
+        }
+    }
+}
+
+impl LineStyle {
+    /// Creates a new line style with default values.
+    pub fn new() -> LineStyle {
+        LineStyle::default()
+    }
+
+    /// Sets the line thickness.
+    ///
+    /// Setting this to 0.0 will not hide the line, rather it’s a special value that tells PDF
+    /// viewers to render the line as 1px regardless of the display size and zoom.
+    pub fn set_thickness(&mut self, thickness: impl Into<Mm>) {
+        self.thickness = thickness.into();
+    }
+
+    /// Sets the line thickness and returns the line style.
+    ///
+    /// Setting this to 0.0 will not hide the line, rather it’s a special value that tells PDF
+    /// viewers to render the line as 1px regardless of the display size and zoom.
+    pub fn with_thickness(mut self, thickness: impl Into<Mm>) -> Self {
+        self.set_thickness(thickness);
+        self
+    }
+
+    /// Returns the line thickness.
+    pub fn thickness(&self) -> Mm {
+        self.thickness
+    }
+
+    /// Sets the line color.
+    pub fn set_color(&mut self, color: Color) {
+        self.color = color;
+    }
+
+    /// Sets the line color and returns the line style.
+    pub fn with_color(mut self, color: Color) -> Self {
+        self.set_color(color);
+        self
+    }
+
+    /// Returns the line color.
+    pub fn color(&self) -> Color {
+        self.color
+    }
+}